add support for new callback
