+++ system/library/spectrocoin/Enum/OrderStatus.php
@@ -6,10 +6,45 @@ if (!defined('DIR_APPLICATION')) {
     die('Access denied.');
 }
 
-enum OrderStatus: int {
-	case New = 1;
-	case Pending = 2;
-	case Paid = 3;
-	case Failed = 4;
-	case Expired = 5;
+enum OrderStatus: string
+{
+    case NEW     = 'NEW';
+    case PENDING = 'PENDING';
+    case PAID    = 'PAID';
+    case FAILED  = 'FAILED';
+    case EXPIRED = 'EXPIRED';
+
+    /**
+     * Map old numeric codes to new enum.
+     */
+    public static function fromCode(int $code): self
+    {
+        return match ($code) {
+            1 => self::NEW,
+            2 => self::PENDING,
+            3 => self::PAID,
+            4 => self::FAILED,
+            5 => self::EXPIRED,
+            default => throw new \InvalidArgumentException("Unknown numeric status code: $code"),
+        };
+    }
+
+    /**
+     * Normalize either an integer (legacy) or a string.
+     */
+    public static function normalize(string|int $raw): self
+    {
+        if (is_int($raw) || ctype_digit((string)$raw)) {
+            return self::fromCode((int)$raw);
+        }
+        $upper = strtoupper((string)$raw);
+        return match ($upper) {
+            'NEW'     => self::NEW,
+            'PENDING' => self::PENDING,
+            'PAID'    => self::PAID,
+            'FAILED'  => self::FAILED,
+            'EXPIRED' => self::EXPIRED,
+            default   => throw new \InvalidArgumentException("Unknown status string: $raw"),
+        };
+    }
 }
